refactor(predict): new predict feed with improved animations
