--- app/components/UI/Predict/components/MarketListContent/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './MarketListContent';
