--- app/components/UI/Predict/components/PredictFeedHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PredictFeedHeader';
