--- app/components/UI/Predict/components/PredictMarketList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PredictMarketList';
