--- app/components/UI/Predict/components/SearchBox/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './SearchBox';
